Reorganize ExperimentRowCard component into bundle directory
Issue #2807
